Create Bridge-to-Polkadot.py
This program simulates first part of a token bridge from VeChain to Polkadot, by sending VET from one wallet to another. 
Sending wallet is a token holder's wallet
Receiving wallet is VeChain part of a 'smart contract' that together with other functions bridges the tokens to a Polkadot version, 
with information about which wallet on Polkadot that has to receive the bridged tokens.